moved dashboard, and dashboard layout into app dir folder
